+++ src/lib.rs
@@ -13,6 +13,7 @@ mod format;
 mod gamma;
 mod gray;
 mod mask;
+mod palette;
 mod raster;
 mod rgb;
 
@@ -24,5 +25,6 @@ pub use crate::gray::{
     Gray, Gray8, Gray16, Gray32, GrayAlpha8, GrayAlpha16, GrayAlpha32
 };
 pub use crate::mask::{Mask, Mask8, Mask16, Mask32};
+pub use crate::palette::Palette;
 pub use crate::raster::{Raster, RasterBuilder, RasterIter, Region};
 pub use crate::rgb::{Rgb, Rgb8, Rgb16, Rgb32, Rgba8, Rgba16, Rgba32};